Add support for OS X, update halide, improve deps.sh

+++ deps.sh
@@ -1,11 +1,22 @@
 #!/bin/bash
 
-cores=$(nproc)
+if [[ "$OSTYPE" == "linux-gnu" ]]; then
+    cores=$(nproc)
+        # ...
+elif [[ "$OSTYPE" == "darwin"* ]]; then
+    cores=$(sysctl -n hw.ncpu)
+        # Mac OSX
+else
+    # Unknown.
+    echo "Unknown OSTYPE: $OSTYPE. Exiting."
+    exit 1
+fi
 
 LOCAL_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 BUILD_DIR=$LOCAL_DIR/thirdparty/build
 DEFAULT_INSTALL_DIR=$LOCAL_DIR/thirdparty/install
 FILES_DIR=$LOCAL_DIR/thirdparty/resources
+PYTHON_VERSION=$(python3 -c 'import sys; print(".".join(map(str, sys.version_info[:2])))')
 
 POSITIONAL=()
 
@@ -76,17 +87,35 @@ echo "(customized by specifying (--prefix <dir>)"
 
 set -- "${POSITIONAL[@]}" # restore positional parameters
 
+if command -v conda list >/dev/null 2>&1; then
+    # Anaconda is installed, so add lib to prefix path for OpenCV to find
+    # PythonLib
+    echo "Detected Anaconda, adding lib path to OpenCV and Caffe build"
+    py_path=$(dirname $(which python))/../lib
+    PY_EXTRA_CMDS="$py_path"
+else
+    PY_EXTRA_CMDS=""
+fi
+
 # Check if we have GPUs by looking for nvidia-smi
-if command -v nvidia-smi; then
+if command -v nvidia-smi >/dev/null 2>&1; then
     HAVE_GPU=true
 else
     HAVE_GPU=false
 fi
 
+
 # Force building with GPU when specified
 if [[ $USE_GPU == true ]]; then
     HAVE_GPU=true
 fi
+
+echo ""
+echo "Configuration:"
+echo "--------------------------------------------------------------"
+echo "Detected Python version: $PYTHON_VERSION"
+echo "GPUs available:          $HAVE_GPU"
+echo ""
 
 # Directories for installed dependencies
 FFMPEG_DIR=$INSTALL_PREFIX
@@ -202,7 +231,7 @@ elif [[ $INSTALL_ALL == false ]]; then
     done
 
     while true; do
-        echo -n "Do you have halide (release_2016_10_25) installed? [y/N]: "
+        echo -n "Do you have halide (release_2018_02_15) installed? [y/N]: "
         read yn
         if [[ $yn == y ]] || [[ $yn == Y ]]; then
             INSTALL_HALIDE=false
@@ -257,13 +286,13 @@ elif [[ $INSTALL_ALL == false ]]; then
         else
             INSTALL_CAFFE=true
             if [[ $HAVE_GPU == true ]]; then
-                echo -n "Do you plan to use GPUs for CNN evaluation? [y/N]: "
+                echo -n "Do you plan to use GPUs for CNN evaluation? [Y/n]: "
                 read yn
-                if [[ $yn == y ]] || [[ $yn == Y ]]; then
-                    USE_GPU=true
+                if [[ $yn == n ]] || [[ $yn == N ]]; then
+                    USE_GPU=false
                     break
                 else
-                    USE_GPU=false
+                    USE_GPU=true
                     break
                 fi
             else
@@ -274,17 +303,36 @@ elif [[ $INSTALL_ALL == false ]]; then
     done
 fi
 
+echo ""
+
 if [[ $INSTALL_FFMPEG == true ]] && [[ ! -f $BUILD_DIR/ffmpeg.done ]] ; then
     echo "Installing ffmpeg 3.3.1..."
+
+    # Determine command string to use
+    if [[ "$OSTYPE" == "linux-gnu" ]]; then
+        # Linux
+        CMDS="--extra-version=0ubuntu0.16.04.1
+              --toolchain=hardened
+              --cc=cc --cxx=g++"
+        # ...
+    elif [[ "$OSTYPE" == "darwin"* ]]; then
+        # Mac OSX
+        CMDS=""
+    fi
+
     # FFMPEG
     cd $BUILD_DIR
     rm -fr ffmpeg
     git clone -b n3.3.1 https://git.ffmpeg.org/ffmpeg.git && cd ffmpeg && \
-    ./configure --prefix=$INSTALL_PREFIX --extra-version=0ubuntu0.16.04.1 \
-                --toolchain=hardened --cc=cc --cxx=g++ --enable-gpl \
+    ./configure --prefix=$INSTALL_PREFIX \
                 --enable-shared --disable-stripping \
-                --disable-decoder=libschroedinger --enable-gnutls \
-                --enable-avresample --enable-libx264 --enable-nonfree && \
+                --disable-decoder=libschroedinger \
+                --enable-avresample \
+                --enable-libx264 \
+                --enable-nonfree \
+                --enable-gpl \
+                --enable-gnutls \
+                $(echo $CMDS) && \
     make -j${cores} && make install && touch $BUILD_DIR/ffmpeg.done \
         || { echo 'Installing ffmpeg failed!' ; exit 1; }
     echo "Done installing ffmpeg 3.3.1"
@@ -293,21 +341,36 @@ fi
 if [[ $INSTALL_OPENCV == true ]] && [[ ! -f $BUILD_DIR/opencv.done ]]; then
     # OpenCV 3.4.0 + OpenCV contrib
     echo "Installing OpenCV 3.4.0..."
+
+    # Determine command string to use
+    if [[ "$OSTYPE" == "linux-gnu" ]]; then
+        # Linux
+        CMDS=""
+        # ...
+    elif [[ "$OSTYPE" == "darwin"* ]]; then
+        # Mac OSX
+        CMDS="-DWITH_CUDA=OFF"
+    fi
+
     cd $BUILD_DIR
     rm -rf opencv opencv_contrib ceres-solver
-    git clone -b 3.4.1 https://github.com/opencv/opencv && \
-        git clone -b 3.4.1  https://github.com/opencv/opencv_contrib && \
-        git clone -b 1.14.0 https://github.com/ceres-solver/ceres-solver && \
-        cd ceres-solver && mkdir build && cd build && \
+    git clone -b 3.4.1 https://github.com/opencv/opencv --depth 1 && \
+        git clone -b 3.4.1  https://github.com/opencv/opencv_contrib \
+            --depth 1 && \
+        git clone -b 1.14.0 https://github.com/ceres-solver/ceres-solver \
+            --depth 1 && \
+        cd ceres-solver && mkdir -p build_cmake && cd build_cmake && \
         cmake .. -DCMAKE_INSTALL_PREFIX=$INSTALL_PREFIX && \
         make install -j$cores && \
         mkdir -p $BUILD_DIR/opencv/build && cd $BUILD_DIR/opencv/build && \
-        cmake -D CMAKE_BUILD_TYPE=Release -D CMAKE_INSTALL_PREFIX=$INSTALL_PREFIX \
+        cmake -D CMAKE_BUILD_TYPE=Release \
+              -D CMAKE_INSTALL_PREFIX=$INSTALL_PREFIX \
               -D BUILD_TESTS=OFF -D BUILD_PERF_TESTS=OFF -D ENABLE_FAST_MATH=1 \
               -D CUDA_FAST_MATH=1 -D WITH_CUBLAS=1 -D WITH_NVCUVID=1 \
               -D BUILD_opencv_rgbd=OFF \
               -D BUILD_opencv_cnn_3dobj=OFF \
               -D OPENCV_EXTRA_MODULES_PATH=$BUILD_DIR/opencv_contrib/modules \
+              $(echo $CMDS) -DCMAKE_PREFIX_PATH=$(echo $PY_EXTRA_CMDS) \
               .. && \
         make install -j$cores && touch $BUILD_DIR/opencv.done \
             || { echo 'Installing OpenCV failed!' ; exit 1; }
@@ -319,7 +382,7 @@ if [[ $INSTALL_PROTOBUF == true ]] && [[ ! -f $BUILD_DIR/protobuf.done ]] ; then
     echo "Installing protobuf 3.4.1..."
     cd $BUILD_DIR
     rm -fr protobuf
-    git clone -b v3.4.1 https://github.com/google/protobuf.git && \
+    git clone -b v3.4.1 https://github.com/google/protobuf.git --depth 1 && \
         cd protobuf && bash ./autogen.sh && \
         ./configure --prefix=$INSTALL_PREFIX && make -j$cores && \
         make install && touch $BUILD_DIR/protobuf.done \
@@ -334,27 +397,74 @@ if [[ $INSTALL_GRPC == true ]] && [[ ! -f $BUILD_DIR/grpc.done ]] ; then
     rm -fr grpc
     git clone -b v1.7.2 https://github.com/grpc/grpc && \
         cd grpc && git submodule update --init --recursive && \
-        make EXTRA_CPPFLAGS=-I$INSTALL_PREFIX/include \
-             EXTRA_LDFLAGS=-L$INSTALL_PREFIX/lib -j$cores && \
-        make install EXTRA_CPPFLAGS=-I$INSTALL_PREFIX/include \
-             EXTRA_LDFLAGS=-L$INSTALL_PREFIX/lib prefix=$INSTALL_PREFIX && \
-        ldconfig -n $INSTALL_PREFIX/lib && \
+        CPPFLAGS=-I$INSTALL_PREFIX/include LDFLAGS=-L$INSTALL_PREFIX/lib make -j$cores && \
+        CPPFLAGS=-I$INSTALL_PREFIX/include LDFLAGS=-L$INSTALL_PREFIX/lib make install prefix=$INSTALL_PREFIX && \
         touch $BUILD_DIR/grpc.done \
             || { echo 'Installing gRPC failed!' ; exit 1; }
+    if [[ "$OSTYPE" == "linux-gnu" ]]; then
+        # Linux
+        ldconfig -n $INSTALL_PREFIX/lib
+    elif [[ "$OSTYPE" == "darwin"* ]]; then
+        # OS X
+        install_name_tool -id "@rpath/libgrpc++_unsecure.dylib" \
+                          $INSTALL_PREFIX/lib/libgrpc++_unsecure.dylib
+        install_name_tool -id "@rpath/libgrpc.dylib" \
+                          $INSTALL_PREFIX/lib/libgrpc.dylib
+        install_name_tool -id "@rpath/libgpr.dylib" \
+                          $INSTALL_PREFIX/lib/libgpr.dylib
+        install_name_tool -change libgpr.dylib @rpath/libgpr.dylib \
+                          $INSTALL_PREFIX/lib/libgrpc++_unsecure.dylib
+        install_name_tool -change libgrpc_unsecure.dylib @rpath/libgrpc_unsecure.dylib \
+                          $INSTALL_PREFIX/lib/libgrpc++_unsecure.dylib
+    fi
     echo "Done installing gRPC 1.7.2"
 fi
 
 if [[ $INSTALL_HALIDE == true ]] && [[ ! -f $BUILD_DIR/halide.done ]] ; then
     # Halide
     echo "Installing Halide..."
+
     cd $BUILD_DIR
     rm -fr Halide
-    git clone -b release_2016_10_25 https://github.com/halide/Halide && \
-        cd Halide && \
-        make distrib -j$cores && \
-        cp -r distrib/* $INSTALL_PREFIX && \
-        touch $BUILD_DIR/halide.done \
-            || { echo 'Installing Halide failed!' ; exit 1; }
+    mkdir Halide
+    cd Halide
+    if [[ "$OSTYPE" == "linux-gnu" ]]; then
+        # If CLANG is not set, we should set it to clang or clang-5.0
+        if [ -z ${CLANG+x} ]; then
+            if command -v clang >/dev/null 2>&1; then
+                export CLANG=clang
+            elif command -v clang-5.0 >/dev/null 2>&1; then
+                export CLANG=clang-5.0
+            fi
+        fi
+        # If LLVM_CONFIG is not set, we should set it to llvm-config or
+        # llvm-config-5.0
+        if [ -z ${LLVM_CONFIG+x} ]; then
+            if command -v llvm-config >/dev/null 2>&1; then
+                export LLVM_CONFIG=llvm-config
+            elif command -v llvm-config-5.0 >/dev/null 2>&1; then
+                export LLVM_CONFIG=llvm-config-5.0
+            fi
+        fi
+        git clone -b release_2018_02_15 https://github.com/halide/Halide --depth 1 && \
+            cd Halide && \
+            make distrib -j$cores && \
+            cp -r distrib/* $INSTALL_PREFIX && \
+            touch $BUILD_DIR/halide.done \
+                || { echo 'Installing Halide failed!' ; exit 1; }
+    elif [[ "$OSTYPE" == "darwin"* ]]; then
+        TAR_NAME=halide-mac-64-trunk-46d8e9e0cdae456489f1eddfd6d829956fc3c843.tgz
+        wget https://github.com/halide/Halide/releases/download/release_2018_02_15/$TAR_NAME && \
+            wget https://raw.githubusercontent.com/halide/Halide/release_2018_02_15/src/Generator.h && \
+            tar -zxf $TAR_NAME && \
+            rm -r halide/bin && \
+            cp Generator.h halide/include && \
+            cp -r halide/* $INSTALL_PREFIX && \
+            cp -r halide/bin/* $INSTALL_PREFIX/lib && \
+            touch $BUILD_DIR/halide.done \
+                || { echo 'Installing Halide failed!' ; exit 1; }
+    fi
+
     echo "Done installing Halide"
 fi
 
@@ -362,7 +472,7 @@ if [[ $INSTALL_PYBIND == true ]] && [[ ! -f $BUILD_DIR/pybind.done ]] ; then
     echo "Installing pybind..."
     cd $BUILD_DIR
     rm -fr pybind11
-    git clone -b v2.2.2 https://github.com/pybind/pybind11 && \
+    git clone -b v2.2.2 https://github.com/pybind/pybind11 --depth 1 && \
         cd pybind11 && \
         mkdir build && cd build && \
         cmake .. -DCMAKE_INSTALL_PREFIX=$INSTALL_PREFIX -DPYBIND11_TEST=Off && \
@@ -377,7 +487,7 @@ if [[ $INSTALL_STOREHOUSE == true ]] && [[ ! -f $BUILD_DIR/storehouse.done ]] ; 
     cd $BUILD_DIR
     rm -fr storehouse
     git clone https://github.com/scanner-research/storehouse && \
-        cd storehouse && git checkout 8d7f6a0853da085dfdf90e4fc3aae286dffe64dd && \
+        cd storehouse && git checkout a7356274d47e11e4e479be3181c21e732f63f60b && \
         cd thirdparty && mkdir build && cd build && \
         cmake .. -DCMAKE_INSTALL_PREFIX=$INSTALL_PREFIX && \
         make -j${cores} && cd ../../ && \
@@ -408,7 +518,7 @@ if [[ $INSTALL_HWANG == true ]] && [[ ! -f $BUILD_DIR/hwang.done ]] ; then
     rm -fr hwang
     git clone https://github.com/scanner-research/hwang && \
         cd hwang && \
-        git checkout acd150d14706b545cf1bace756e5c6b684187c39 && \
+        git checkout 45abc5203b9b32d341351430bbea9c7050f3e037 && \
         bash ./deps.sh -a \
              --with-ffmpeg $INSTALL_PREFIX \
              --with-protobuf $INSTALL_PREFIX \
@@ -434,21 +544,22 @@ if [[ $INSTALL_TINYTOML == true ]] && [[ ! -f $BUILD_DIR/tinytoml.done ]]; then
 fi
 
 if [[ $INSTALL_CAFFE == true ]] && [[ $USE_GPU == false ]] && \
-       [[ ! -f $BUILD_DIR/caffe.done ]]; then
+       [[ "$OSTYPE" == "linux-gnu" ]] && [[ ! -f $BUILD_DIR/caffe.done ]]; then
     # Intel Caffe 1.0.6
     cd $BUILD_DIR
     rm -fr caffe
     # Use more recent mkldnn commit to fix gcc bug
-    git clone -b 1.0.6 https://github.com/intel/caffe && \
+    git clone -b 1.0.6 https://github.com/intel/caffe --depth 1 && \
         cd caffe && \
         cp $FILES_DIR/caffe/Makefile.config Makefile.config && \
         rm mkldnn.commit && \
         echo "2604f435da7bb9f1896ae37200d91734adfdba9c" > mkldnn.commit && \
         mkdir build && cd build && \
         cmake -DCMAKE_INSTALL_PREFIX=$INSTALL_PREFIX \
-              -DCMAKE_PREFIX_PATH=$INSTALL_PREFIX \
+              -DCMAKE_PREFIX_PATH="$INSTALL_PREFIX;$PY_EXTRA_CMDS" \
               -DCPU_ONLY=ON \
-              -DOpenCV_DIR=$INSTALL_PREFIX \
+              -DOpenCV_DIR=$OPENCV_DIR \
+              -DBUILD_python=OFF \
               -Dpython_version=3 \
               -DBLAS=mkl \
               .. && \
@@ -461,19 +572,31 @@ if [[ $INSTALL_CAFFE == true ]] && [[ $USE_GPU == false ]] && \
             || { echo 'Installing caffe failed!' ; exit 1; }
 fi
 
-if [[ $INSTALL_CAFFE == true ]] && [[ $USE_GPU == true ]] && \
+#if [[ $INSTALL_CAFFE == true ]] && 
+if [[ $INSTALL_CAFFE == true ]] && \
+       ([[ $USE_GPU == true ]] ||
+        [[ "$OSTYPE" == "darwin"* ]]) && \
        [[ ! -f $BUILD_DIR/caffe.done ]]; then
     cd $BUILD_DIR
     # Intel MKL
-    rm -fr mkl
-    mkdir mkl && \
-        cd mkl && \
-        wget http://registrationcenter-download.intel.com/akdlm/irc_nas/tec/12414/l_mkl_2018.1.163.tgz && \
-        tar -zxf l_mkl_2018.1.163.tgz && \
-        cp $FILES_DIR/mkl/silent.cfg silent.cfg && \
-        echo "PSET_INSTALL_DIR=$INSTALL_PREFIX/intel" >> silent.cfg && \
-        cd l_mkl_2018.1.163 && \
-        bash install.sh --cli-mode --silent ../silent.cfg
+
+    if [[ "$OSTYPE" == "linux-gnu" ]]; then
+        rm -fr mkl
+        mkdir mkl && \
+            cd mkl && \
+            wget http://registrationcenter-download.intel.com/akdlm/irc_nas/tec/12414/l_mkl_2018.1.163.tgz && \
+            tar -zxf l_mkl_2018.1.163.tgz && \
+            cp $FILES_DIR/mkl/silent.cfg silent.cfg && \
+            echo "PSET_INSTALL_DIR=$INSTALL_PREFIX/intel" >> silent.cfg && \
+            cd l_mkl_2018.1.163 && \
+            bash install.sh --cli-mode --silent ../silent.cfg
+    fi
+
+    if [[ $USE_GPU == true ]]; then
+        CPU_ONLY=OFF
+    else
+        CPU_ONLY=ON
+    fi
 
     cd $BUILD_DIR
     # Caffe rc5
@@ -486,7 +609,9 @@ if [[ $INSTALL_CAFFE == true ]] && [[ $USE_GPU == true ]] && \
         cmake -DCMAKE_INSTALL_PREFIX=$INSTALL_PREFIX \
               -DCMAKE_PREFIX_PATH=$INSTALL_PREFIX \
               -DINTEL_ROOT=$INSTALL_PREFIX/intel \
+              -DCPU_ONLY=$CPU_ONLY \
               -DBLAS=mkl \
+              -DBUILD_python=OFF \
               -Dpython_version=3 \
               -DCUDA_ARCH_NAME="Manual" \
               -DCUDA_ARCH_BIN="30 35 50 60 61" \
@@ -502,7 +627,7 @@ fi
 if [[ $INSTALL_OPENPOSE == true ]] && [[ $HAVE_GPU == true ]] && [[ ! -f $BUILD_DIR/openpose.done ]]; then
     cd $BUILD_DIR
     rm -rf openpose
-    git clone -b v1.2.0 https://github.com/CMU-Perceptual-Computing-Lab/openpose && \
+    git clone -b v1.2.0 https://github.com/CMU-Perceptual-Computing-Lab/openpose --depth 1 && \
         cd openpose && mkdir build && cd build && \
         cmake -D CMAKE_INSTALL_PREFIX=$INSTALL_PREFIX \
               -D CMAKE_PREFIX_PATH=$INSTALL_PREFIX \
@@ -547,7 +672,7 @@ echo -n "add $INSTALL_PREFIX/bin to your PATH so the installed "
 echo -n "dependencies can be found! "
 echo "e.g. export LD_LIBRARY_PATH=$INSTALL_PREFIX/lib:\$LD_LIBRARY_PATH"
 if [[ $INSTALL_OPENCV == true ]]; then
-    echo "Add $INSTALL_PREFIX/lib/python2.7/dist-packages to your PYTHONPATH to use OpenCV from Python"
+    echo "Add $INSTALL_PREFIX/lib/python$PYTHON_VERSION/dist-packages to your PYTHONPATH to use OpenCV from Python"
 fi
 if [[ $INSTALL_CAFFE_CPU == true ]] || [[ $INSTALL_CAFFE_GPU == true ]]; then
     echo "Add $INSTALL_PREFIX/python to your PYTHONPATH to use Caffe from Python"
